Move KMS classes to file_format_options.rs and encryption.rs
