+++ crates/deltalake/examples/basic_operations_encryption.rs
@@ -1,78 +1,39 @@
-use async_trait::async_trait;
 use deltalake::arrow::{
     array::{Int32Array, StringArray, TimestampMicrosecondArray},
     datatypes::{
-        DataType as ArrowDataType, Field, Schema, Schema as ArrowSchema, SchemaRef, TimeUnit,
+        DataType as ArrowDataType, Field, Schema, Schema as ArrowSchema, TimeUnit,
     },
     record_batch::RecordBatch,
 };
 use deltalake::datafusion::{
     assert_batches_sorted_eq,
-    catalog::Session,
-    common::{extensions_options, DataFusionError},
-    config::{
-        ConfigEntry, ConfigFileType, EncryptionFactoryOptions, ExtensionOptions,
-        TableOptions, TableParquetOptions,
-    },
+    config::{ConfigFileType, TableOptions, TableParquetOptions},
     dataframe::DataFrame,
-    execution::parquet_encryption::EncryptionFactory,
     logical_expr::{col, lit},
     prelude::SessionContext,
 };
 use deltalake::kernel::{DataType, PrimitiveType, StructField};
 use deltalake::operations::collect_sendable_stream;
-use deltalake::parquet::encryption::{
-    decrypt::FileDecryptionProperties, encrypt::FileEncryptionProperties,
-};
+use deltalake::parquet::encryption::decrypt::FileDecryptionProperties;
 use deltalake::{arrow, parquet, DeltaOps};
-use deltalake_core::operations::encryption::TableEncryption;
+use deltalake_core::operations::encryption::{
+    KmsEncryptionFactory, KmsEncryptionFactoryOptions, TableEncryption,
+};
 use deltalake_core::table::file_format_options::{
-    FileFormatOptions, FileFormatRef, KMSWriterPropertiesFactory, SimpleFileFormatOptions,
-    WriterPropertiesFactory,
+    FileFormatRef, KmsFileFormatOptions, SimpleFileFormatOptions,
 };
 use deltalake_core::{
-    checkpoints, datafusion::common::test_util::format_batches, operations::optimize::OptimizeType,
-    DeltaResult, DeltaTable, DeltaTableError,
+    datafusion::common::test_util::format_batches, operations::optimize::OptimizeType, DeltaTable,
+    DeltaTableError,
 };
 use parquet_key_management::{
     crypto_factory::{CryptoFactory, DecryptionConfiguration, EncryptionConfiguration},
     kms::KmsConnectionConfig,
     test_kms::TestKmsClientFactory,
 };
-use std::{
-    any::Any,
-    collections::HashMap,
-    fmt::{Debug, Formatter},
-    fs,
-    path::PathBuf,
-    sync::Arc,
-    time::Duration,
-};
+use std::{fs, sync::Arc};
 use tempfile::TempDir;
 use url::Url;
-use uuid::Uuid;
-
-#[tokio::main(flavor = "current_thread")]
-async fn main() -> Result<(), DeltaTableError> {
-    println!("====================");
-    println!("Begin Plain encryption test");
-    println!("");
-    let file_format_options = plain_crypto_format()?;
-    round_trip_test(file_format_options).await?;
-    println!("End Plain encryption test");
-    println!("====================");
-
-    println!("\n\n");
-    println!("====================");
-    println!("Begin KMS encryption test");
-    println!("");
-    let file_format_options = kms_crypto_format()?;
-    round_trip_test(file_format_options).await?;
-    println!("End KMS encryption test");
-    println!("====================");
-
-    Ok(())
-}
 
 fn get_table_columns() -> Vec<StructField> {
     vec![
@@ -374,329 +335,22 @@ fn kms_crypto_format() -> Result<FileFormatRef, DeltaTableError> {
     Ok(file_format_options)
 }
 
+#[tokio::main(flavor = "current_thread")]
+async fn main() -> Result<(), DeltaTableError> {
+    println!("====================");
+    println!("Begin Plain encryption test");
+    let file_format_options = plain_crypto_format()?;
+    round_trip_test(file_format_options).await?;
+    println!("End Plain encryption test");
+    println!("====================");
 
-// -------------------------------------------------------------------------------------------------
-// An example of FileFormatOptions that uses the KMS encryption factory.
-// -------------------------------------------------------------------------------------------------
+    println!("\n\n");
+    println!("====================");
+    println!("Begin KMS encryption test");
+    let file_format_options = kms_crypto_format()?;
+    round_trip_test(file_format_options).await?;
+    println!("End KMS encryption test");
+    println!("====================");
 
-struct KmsFileFormatOptions {
-    table_encryption: TableEncryption,
-    writer_properties_factory: Arc<dyn WriterPropertiesFactory>,
-    encryption_factory_id: String,
-}
-
-impl KmsFileFormatOptions {
-    fn new(table_encryption: TableEncryption) -> Self {
-        let encryption_factory_id = format!("delta-{}", Uuid::new_v4().to_string());
-        let writer_properties_factory = Arc::new(KMSWriterPropertiesFactory::with_encryption(
-            table_encryption.clone(),
-        ));
-        Self {
-            table_encryption,
-            writer_properties_factory,
-            encryption_factory_id,
-        }
-    }
-}
-
-impl Debug for KmsFileFormatOptions {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("KmsFileFormatOptions")
-            .finish_non_exhaustive()
-    }
-}
-
-impl FileFormatOptions for KmsFileFormatOptions {
-    fn table_options(&self) -> TableOptions {
-        let mut table_options = TableOptions::default();
-        table_options.parquet.crypto.factory_id = Some(self.encryption_factory_id.clone());
-        table_options.parquet.crypto.factory_options =
-            self.table_encryption.configuration().clone();
-        table_options
-    }
-
-    fn writer_properties_factory(&self) -> Arc<dyn WriterPropertiesFactory> {
-        Arc::clone(&self.writer_properties_factory)
-    }
-
-    fn update_session(&self, session: &dyn Session) -> DeltaResult<()> {
-        // Ensure DataFusion has the encryption factory registered
-        session.runtime_env().register_parquet_encryption_factory(
-            &self.encryption_factory_id,
-            Arc::clone(self.table_encryption.encryption_factory()),
-        );
-        Ok(())
-    }
-}
-
-// -------------------------------------------------------------------------------------------------
-// Below code should all be part of parquet-key-management-rs.
-// See PR to add DataFusion integration:
-// https://github.com/G-Research/parquet-key-management-rs/pull/20
-// -------------------------------------------------------------------------------------------------
-
-pub struct KmsEncryptionFactory {
-    crypto_factory: CryptoFactory,
-    kms_connection_config: Arc<KmsConnectionConfig>,
-}
-
-impl KmsEncryptionFactory {
-    /// Create a new [`KmsEncryptionFactory`] with the provided [`CryptoFactory`] and [`KmsConnectionConfig`].
-    pub fn new(
-        crypto_factory: CryptoFactory,
-        kms_connection_config: Arc<KmsConnectionConfig>,
-    ) -> Self {
-        Self {
-            crypto_factory,
-            kms_connection_config,
-        }
-    }
-}
-
-#[async_trait]
-impl EncryptionFactory for KmsEncryptionFactory {
-    async fn get_file_encryption_properties(
-        &self,
-        config: &EncryptionFactoryOptions,
-        _schema: &SchemaRef,
-        _file_path: &object_store::path::Path,
-    ) -> deltalake::datafusion::common::Result<Option<FileEncryptionProperties>> {
-        let encryption_configuration = build_encryption_configuration(config)?;
-        Ok(Some(self.crypto_factory.file_encryption_properties(
-            Arc::clone(&self.kms_connection_config),
-            &encryption_configuration,
-        )?))
-    }
-
-    async fn get_file_decryption_properties(
-        &self,
-        config: &EncryptionFactoryOptions,
-        _file_path: &object_store::path::Path,
-    ) -> deltalake::datafusion::common::Result<Option<FileDecryptionProperties>> {
-        let decryption_configuration = build_decryption_configuration(config)?;
-        Ok(Some(self.crypto_factory.file_decryption_properties(
-            Arc::clone(&self.kms_connection_config),
-            decryption_configuration,
-        )?))
-    }
-}
-
-impl std::fmt::Debug for KmsEncryptionFactory {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("KmsEncryptionFactory")
-            .finish_non_exhaustive()
-    }
-}
-
-/// DataFusion compatible options used to configure generation of encryption and decryption
-/// properties for a Table when using a [`KmsEncryptionFactory`].
-#[derive(Clone, Debug, Default, PartialEq)]
-pub struct KmsEncryptionFactoryOptions {
-    /// The configuration to use when writing encrypted Parquet
-    pub encryption: EncryptionOptions,
-    /// The configuration to use when reading encrypted Parquet
-    pub decryption: DecryptionOptions,
-}
-
-impl KmsEncryptionFactoryOptions {
-    /// Create a new [`KmsEncryptionFactoryOptions `] from encryption and decryption configurations
-    pub fn new(
-        encryption_config: EncryptionConfiguration,
-        decryption_config: DecryptionConfiguration,
-    ) -> Self {
-        Self {
-            encryption: encryption_config.into(),
-            decryption: decryption_config.into(),
-        }
-    }
-}
-
-extensions_options! {
-    /// DataFusion compatible configuration options related to file encryption
-    pub struct EncryptionOptions {
-        /// Master key identifier for footer key encryption
-        pub footer_key_id: String, default = "".to_owned()
-        /// List of master key ids and the columns to be encrypted with each key,
-        /// formatted like "columnKeyId1:column1,column2;columnKeyId2:column3"
-        pub column_key_ids: String, default = "".to_owned()
-        /// Whether to write footer metadata unencrypted
-        pub plaintext_footer: bool, default = false
-        /// Whether to encrypt data encryption keys with key encryption keys, before wrapping with the master key
-        pub double_wrapping: bool, default = true
-        /// How long in seconds to cache objects used during encryption
-        pub cache_lifetime_s: Option<u64>, default = None
-        /// Whether to store encryption key material inside Parquet files
-        pub internal_key_material: bool, default = true
-        /// Length of data encryption keys to generate
-        pub data_key_length_bits: u64, default = 128
-    }
-}
-
-extensions_options! {
-    /// DataFusion compatible configuration options related to file decryption
-    pub struct DecryptionOptions {
-        /// How long in seconds to cache objects used during decryption
-        pub cache_lifetime_s: Option<u64>, default = None
-    }
-}
-
-// Manually implement PartialEq as using #[derive] isn't compatible with extensions_options macro
-impl PartialEq for EncryptionOptions {
-    fn eq(&self, other: &Self) -> bool {
-        self.footer_key_id == other.footer_key_id
-            && self.column_key_ids == other.column_key_ids
-            && self.plaintext_footer == other.plaintext_footer
-            && self.double_wrapping == other.double_wrapping
-            && self.cache_lifetime_s == other.cache_lifetime_s
-            && self.internal_key_material == other.internal_key_material
-            && self.data_key_length_bits == other.data_key_length_bits
-    }
-}
-
-impl PartialEq for DecryptionOptions {
-    fn eq(&self, other: &Self) -> bool {
-        self.cache_lifetime_s == other.cache_lifetime_s
-    }
-}
-
-impl ExtensionOptions for KmsEncryptionFactoryOptions {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn as_any_mut(&mut self) -> &mut dyn Any {
-        self
-    }
-
-    fn cloned(&self) -> Box<dyn ExtensionOptions> {
-        Box::new(self.clone())
-    }
-
-    fn set(&mut self, key: &str, value: &str) -> deltalake::datafusion::common::Result<()> {
-        let (parent, key) = key.split_once('.').ok_or_else(|| {
-            DataFusionError::Configuration(format!(
-                "Invalid configuration key for KMS encryption: {key}"
-            ))
-        })?;
-        match parent {
-            "decryption" => ExtensionOptions::set(&mut self.decryption, key, value),
-            "encryption" => ExtensionOptions::set(&mut self.encryption, key, value),
-            _ => Err(DataFusionError::Configuration(format!(
-                "Invalid configuration key for KMS encryption: {parent}"
-            ))),
-        }
-    }
-
-    fn entries(&self) -> Vec<ConfigEntry> {
-        let mut decryption_entries = self.decryption.entries();
-        for entry in decryption_entries.iter_mut() {
-            entry.key = format!("decryption.{}", entry.key);
-        }
-        let mut encryption_entries = self.encryption.entries();
-        for entry in encryption_entries.iter_mut() {
-            entry.key = format!("encryption.{}", entry.key);
-        }
-        decryption_entries.append(&mut encryption_entries);
-        decryption_entries
-    }
-}
-
-impl From<EncryptionConfiguration> for EncryptionOptions {
-    fn from(config: EncryptionConfiguration) -> Self {
-        EncryptionOptions {
-            footer_key_id: config.footer_key_id().to_owned(),
-            column_key_ids: serialize_column_keys(config.column_key_ids()),
-            plaintext_footer: config.plaintext_footer(),
-            double_wrapping: config.double_wrapping(),
-            cache_lifetime_s: config.cache_lifetime().map(|lifetime| lifetime.as_secs()),
-            internal_key_material: config.internal_key_material(),
-            data_key_length_bits: config.data_key_length_bits() as u64,
-        }
-    }
-}
-
-impl TryInto<EncryptionConfiguration> for EncryptionOptions {
-    type Error = DataFusionError;
-
-    fn try_into(self) -> Result<EncryptionConfiguration, Self::Error> {
-        let mut builder = EncryptionConfiguration::builder(self.footer_key_id.clone())
-            .set_double_wrapping(self.double_wrapping)
-            .set_plaintext_footer(self.plaintext_footer)
-            .set_cache_lifetime(self.cache_lifetime_s.map(Duration::from_secs));
-        let column_keys = deserialize_column_keys(&self.column_key_ids)?;
-        for (key, value) in column_keys.into_iter() {
-            builder = builder.add_column_key(key, value);
-        }
-        Ok(builder.build()?)
-    }
-}
-
-impl From<DecryptionConfiguration> for DecryptionOptions {
-    fn from(config: DecryptionConfiguration) -> Self {
-        DecryptionOptions {
-            cache_lifetime_s: config.cache_lifetime().map(|lifetime| lifetime.as_secs()),
-        }
-    }
-}
-
-impl TryInto<DecryptionConfiguration> for DecryptionOptions {
-    type Error = DataFusionError;
-
-    fn try_into(self) -> Result<DecryptionConfiguration, Self::Error> {
-        Ok(DecryptionConfiguration::builder()
-            .set_cache_lifetime(self.cache_lifetime_s.map(Duration::from_secs))
-            .build())
-    }
-}
-
-fn serialize_column_keys(column_key_ids: &HashMap<String, Vec<String>>) -> String {
-    let mut result = String::new();
-    let mut first = true;
-    for (key_id, columns) in column_key_ids.iter() {
-        if !first {
-            result.push(';');
-        }
-        result.push_str(key_id);
-        result.push(':');
-        result.push_str(&columns.join(","));
-        first = false;
-    }
-    result
-}
-
-fn deserialize_column_keys(
-    column_key_ids: &str,
-) -> deltalake::datafusion::common::Result<HashMap<String, Vec<String>>> {
-    let mut keys = HashMap::new();
-    for key_with_cols in column_key_ids.split(';') {
-        if key_with_cols.is_empty() {
-            continue;
-        }
-        let (key_id, cols) = key_with_cols.split_once(':').ok_or_else(|| {
-            DataFusionError::Configuration(format!(
-                "Invalid column_key_ids format in encryption configuration: '{column_key_ids}'"
-            ))
-        })?;
-        let cols = cols
-            .split(',')
-            .filter(|s| !s.is_empty())
-            .map(String::from)
-            .collect();
-        keys.insert(key_id.to_owned(), cols);
-    }
-    Ok(keys)
-}
-
-fn build_decryption_configuration(
-    options: &EncryptionFactoryOptions,
-) -> deltalake::datafusion::common::Result<DecryptionConfiguration> {
-    let kms_options: KmsEncryptionFactoryOptions = options.to_extension_options()?;
-    kms_options.decryption.try_into()
-}
-
-fn build_encryption_configuration(
-    options: &EncryptionFactoryOptions,
-) -> deltalake::datafusion::common::Result<EncryptionConfiguration> {
-    let kms_options: KmsEncryptionFactoryOptions = options.to_extension_options()?;
-    kms_options.encryption.try_into()
+    Ok(())
 }
